+++ packages/expo-image/ios/ImageView.swift
@@ -119,27 +119,12 @@ public final class ImageView: ExpoView {
     if sdImageView.image == nil {
       sdImageView.contentMode = contentFit.toContentMode()
     }
-    var context = SDWebImageContext()
+    var context = createSDWebImageContext(forSource: source, cachePolicy: cachePolicy)
 
     // Cancel currently running load requests.
     cancelPendingOperation()
 
-    // Modify URL request to add headers.
-    if let headers = source.headers {
-      context[SDWebImageContextOption.downloadRequestModifier] = SDWebImageDownloaderRequestModifier(headers: headers)
-    }
-
-    context[.cacheKeyFilter] = createCacheKeyFilter(source.cacheKey)
     context[.imageTransformer] = createTransformPipeline()
-
-    // Tell SDWebImage to use our own class for animated formats,
-    // which has better compatibility with the UIImage and fixes issues with the image duration.
-    context[.animatedImageClass] = AnimatedImage.self
-
-    // Assets from the bundler have `scale` prop which needs to be passed to the context,
-    // otherwise they would be saved in cache with scale = 1.0 which may result in
-    // incorrectly rendered images for resize modes that don't scale (`center` and `repeat`).
-    context[.imageScaleFactor] = source.scale
 
     // It seems that `UIImageView` can't tint some vector graphics. If the `tintColor` prop is specified,
     // we tell the SVG coder to decode to a bitmap instead. This will become useless when we switch to SVGNative coder.
@@ -153,21 +138,7 @@ public final class ImageView: ExpoView {
       ]
     }
 
-    if source.isCachingAllowed {
-      let sdCacheType = cachePolicy.toSdCacheType().rawValue
-      context[.originalQueryCacheType] = sdCacheType
-      context[.originalStoreCacheType] = sdCacheType
-    } else {
-      context[.originalQueryCacheType] = SDImageCacheType.none.rawValue
-      context[.originalStoreCacheType] = SDImageCacheType.none.rawValue
-    }
-    // Set which cache can be used to query and store the downloaded image.
-    // We want to store only original images (without transformations).
-    context[.queryCacheType] = SDImageCacheType.none.rawValue
-    context[.storeCacheType] = SDImageCacheType.none.rawValue
-
-    // Some loaders (e.g. blurhash) need access to the source and the screen scale.
-    context[ImageView.contextSourceKey] = source
+    // Some loaders (e.g. PhotoLibraryAssetLoader) may need to know the screen scale.
     context[ImageView.screenScaleKey] = screenScale
 
     // Do it here so we don't waste resources trying to fetch from a remote URL
@@ -321,22 +292,14 @@ public final class ImageView: ExpoView {
     guard let placeholder = bestPlaceholder, isViewEmpty || !hasAnySource else {
       return
     }
-    var context = SDWebImageContext()
-    let isPlaceholderHash = placeholder.isBlurhash || placeholder.isThumbhash
-
-    context[.imageScaleFactor] = placeholder.scale
-    context[.cacheKeyFilter] = createCacheKeyFilter(placeholder.cacheKey)
-    context[.animatedImageClass] = AnimatedImage.self
 
     // Cache placeholders on the disk. Should we let the user choose whether
     // to cache them or apply the same policy as with the proper image?
     // Basically they are also cached in memory as the `placeholderImage` property,
     // so just `disk` policy sounds like a good idea.
-    context[.queryCacheType] = SDImageCacheType.disk.rawValue
-    context[.storeCacheType] = SDImageCacheType.disk.rawValue
+    var context = createSDWebImageContext(forSource: placeholder, cachePolicy: .disk)
 
-    // Some loaders (e.g. blurhash) need access to the source.
-    context[ImageView.contextSourceKey] = placeholder
+    let isPlaceholderHash = placeholder.isBlurhash || placeholder.isThumbhash
 
     imageManager.loadImage(with: placeholder.uri, context: context, progress: nil) { [weak self] placeholder, _, _, _, finished, _ in
       guard let self = self, let placeholder = placeholder, finished else {
@@ -388,7 +351,7 @@ public final class ImageView: ExpoView {
     sdImageView.layer.frame.origin = offset
   }
 
-  private func renderImage(_ image: UIImage?) {
+  internal func renderImage(_ image: UIImage?) {
     if let transition = transition, transition.duration > 0 {
       let options = transition.toAnimationOptions()
       let seconds = transition.duration / 1000