+++ packages/expo-image/CHANGELOG.md
@@ -8,6 +8,7 @@
 
 - Add basic React Server Component support. ([#29869](https://github.com/expo/expo/pull/29869) by [@EvanBacon](https://github.com/EvanBacon))
 - [iOS] Added `Image.loadAsync` API. ([#25079](https://github.com/expo/expo/pull/25079) by [@tsapeta](https://github.com/tsapeta))
+- [iOS] Added support for rendering shared image refs.
 
 ### 🐛 Bug fixes